+++ includes/payment-gateways/pro/class-gpos-esnekpos.php
@@ -6,9 +6,9 @@
  */
 
 /**
- * GPOS_Esnek_Pos sınıfı.
+ * GPOS_Esnekpos sınıfı.
  */
-class GPOS_Esnek_Pos extends GPOS_Gateway {
+class GPOS_Esnekpos extends GPOS_Gateway {
 
 	/**
 	 * Ödeme geçidi benzersiz kimliği
@@ -53,12 +53,44 @@ class GPOS_Esnek_Pos extends GPOS_Gateway {
 	public $logo = GPOS_ASSETS_DIR_URL . '/images/logo/esnekpos.png';
 
 	/**
+	 * Desteklenen özellikler
+	 *
+	 * @var array $supports
+	 */
+	public $supports = array( 'threed', 'refund', 'installment_api' );
+
+	/**
+	 * Firma müşteri panel bilgisi
+	 *
+	 * @var string $merchant_panel
+	 */
+	public $merchant_panel = 'https://pos.esnekpos.com/Login.aspx';
+
+	/**
+	 * Desteklenen para birimleri
+	 *
+	 * @var array $currencies
+	 */
+	public $currencies = array( 'TRY', 'EUR', 'USD' );
+
+	/**
 	 * Ödeme için gerekli alanların tanımı
 	 *
 	 * @return array
 	 */
 	public function get_payment_fields() : array {
-		return array();
+		return array(
+			array(
+				'type'  => 'text',
+				'label' => __( 'Merchant', 'gurmepos' ),
+				'model' => 'merchant',
+			),
+			array(
+				'type'  => 'text',
+				'label' => __( 'Merchant Key', 'gurmepos' ),
+				'model' => 'merchant_key',
+			),
+		);
 	}
 
 	/**
@@ -67,7 +99,16 @@ class GPOS_Esnek_Pos extends GPOS_Gateway {
 	 * @return array
 	 */
 	public function get_test_credit_cards() : array {
-		return array();
+		return array(
+			array(
+				'type'         => 'Paraf',
+				'bin'          => '9792 1000 0000 0001',
+				'expiry_year'  => '2026',
+				'expiry_month' => '12',
+				'cvv'          => '000',
+				'secure'       => '',
+			),
+		);
 	}
 
 }